requirements.txt
streamlit
numpy
pandas
matplotlib 
plotly.express
seaborn